Fix update loop on pnpm 11
